geolocation-API/getCurrentPosition_permission_allow.https.html

<!DOCTYPE HTML>
<meta charset='utf-8'>
<title>Geolocation Test: getCurrentPosition location access allowed</title>
<link rel='help' href='http://www.w3.org/TR/geolocation-API/#get-current-position'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='/resources/testdriver.js'></script>
<script src='/resources/testdriver-vendor.js'></script>
<script src='support.js'></script>
<script>
// Rewrite http://dev.w3.org/geo/api/test-suite/t.html?00002
promise_test(async t => {
  await test_driver.set_permission({name: 'geolocation'}, 'granted');

  let hasMethodReturned = false;

  return new Promise(resolve => {
    geo.getCurrentPosition(resolve, resolve);
    hasMethodReturned = true;
  }).then(() => {
    assert_true(hasMethodReturned);
  }).catch(err => {
    assert_true(hasMethodReturned);
    assert_false(isUsingPreemptivePermission);
    assert_equals(err.code, err.POSITION_UNAVAILABLE, errorToString(err));
  });
}, 'User allows access, check that success callback is called or error callback is called with correct code.');
</script>
